codebase refactoring #138
